Temporarily update `rainglow` theme github ref
The official repo has an issue which causes the Netlify build to
fail (see https://app.netlify.com/sites/zxul767/deploys/614ffccb5be2759b3b6499c0)

There's a PR to address the issue (https://github.com/rainglow/vscode/pull/38),
but it hasn't been looked at for over a month, so in the meantime I've
decided to reference the fork that contains the fix to unblock myself.

I will revert to the official repo once the PR is merged.