Trim "A Fundamental Developer Skill" post
After implementing the suggested reading time feature, I realized this
post is unnecessarily long (at almost 20 minutes). Part of the reason is
that the second part went into an area that probably requires its own
post. In this commit I've moved that content to another file that I may
turn into a separate post in the future.